siamese_vggface/requirements.txt
torchvision==0.2.2
numpy==1.19.2
pandas==1.2.3
tqdm==4.59.0
torch==1.8.0
Pillow==8.1.2
scikit_learn==0.24.1
